--- src/org-pending/services/index.ts
@@ -1,5 +1,0 @@
-import { PendingOrganisationService } from './pending-organisation.service';
-
-export const services: any[] = [PendingOrganisationService];
-
-export * from './pending-organisation.service';
