+++ src/org-pending/store/effects/org-pending.effects.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Actions, Effect, ofType } from '@ngrx/effects';
-import { PendingOrganisationService } from 'src/org-pending/services/pending-organisation.service';
+import { PendingOrganisationService } from 'src/org-manager/services/pending-organisation.service';
 import * as pendingOrgActions from '../actions/org-pending.actions';
 import { map, switchMap, catchError } from 'rxjs/operators';
 import { of } from 'rxjs';
